fix(datalake): drop testing module-reload bootstrap; relocate benchmark-registration tests
