+++ mindtrace/datalake/mindtrace/datalake/testing/__init__.py
@@ -7,14 +7,12 @@ Use ``register_benchmark_suites`` directly or discover it through the
 from __future__ import annotations
 
 from mindtrace.core import TestRunner
-from mindtrace.datalake.testing.bootstrap import prioritize_wheel_datalake_sources
 
 
 def register_benchmark_suites(*, runner: TestRunner | None = None, replace: bool = True) -> None:
     """Register datalake benchmark suites on ``runner`` or the default runner."""
 
     target = runner or TestRunner.default()
-    prioritize_wheel_datalake_sources()
 
     from mindtrace.datalake.testing.suites.collection_item import DatalakeCollectionItemSuite
     from mindtrace.datalake.testing.suites.create_asset import DatalakeCreateAssetFromObjectSuite
