+++ tests/unit/mindtrace/datalake/test_datalake_benchmark_registration.py
@@ -1,4 +1,4 @@
-"""Embedded benchmark modules expose package-level registration hooks and schemas."""
+"""``mindtrace.datalake.testing`` registers its embedded benchmark suites and schemas."""
 
 from __future__ import annotations
 
@@ -17,47 +17,6 @@ def _assert_suite_schema_contract(suite_schema: object, *, suite_id: str) -> Non
     assert resource_json_schema is not None
     assert resource_json_schema["type"] == "object"
     assert "properties" in resource_json_schema
-
-
-def test_registry_testing_registers_expected_ids_and_schemas() -> None:
-    import mindtrace.registry.testing as rt
-    from mindtrace.core import TestRunner
-
-    TestRunner.clear_registry()
-    rt.register_benchmark_suites()
-
-    ids = sorted(TestRunner.registered_suites())
-    expected = {
-        "registry.smoke.local_crud",
-        "registry.stress.write_ceiling",
-        "registry.stress.read_ceiling",
-        "registry.stress.mixed_rw",
-        "registry.stress.version_churn",
-    }
-    assert expected.issubset(ids)
-
-    for suite_id in expected:
-        _assert_suite_schema_contract(TestRunner.get_suite_schema(suite_id), suite_id=suite_id)
-
-
-def test_database_testing_registers_expected_ids_and_schemas() -> None:
-    import mindtrace.database.testing as dbt
-    from mindtrace.core import TestRunner
-
-    TestRunner.clear_registry()
-    dbt.register_benchmark_suites()
-
-    ids = sorted(TestRunner.registered_suites())
-    expected = {
-        "database.smoke.mongo_crud",
-        "database.stress.mongo_insert_ceiling",
-        "database.stress.mongo_read_ceiling",
-        "database.stress.mongo_update_ceiling",
-    }
-    assert expected.issubset(ids)
-
-    for suite_id in expected:
-        _assert_suite_schema_contract(TestRunner.get_suite_schema(suite_id), suite_id=suite_id)
 
 
 def test_datalake_testing_registers_expected_ids_and_schemas() -> None:
